Fit a Low Rank Gaussian Process Regression (LRGPR) / Linear Mixed Model (LMM) for many markers

Fit LRGPR/LMM models that account for covariance in response values, but where the scale of the covariance is unknown. It returns p-values from a Wald test equivalent to the results of using lrgpr and wald, but is designed to analyze thousands of markers in a single function call.

Usage

lrgprApply(
  formula,
  features,
  decomp,
  terms = NULL,
  rank = max(ncol(decomp$u), ncol(decomp$vectors)),
  map = NULL,
  distance = NULL,
  dcmp_features = NULL,
  W_til = NULL,
  scale = TRUE,
  delta = NULL,
  reEstimateDelta = FALSE,
  nthreads = detectCores(logical = TRUE),
  verbose = FALSE,
  progress = TRUE,
  cincl = c(),
  cexcl = c()
)

Arguments

formula

standard linear modeling syntax as used in `lm'. SNP is a place holder for the each successive column of features

features

a matrix where the statistical model is evaluated with SNP if formula replace by each column successively

decomp

eigen-decomposition produced from eigen(K), where K is the covariance matrix. Or singular value decomposition svd(features[,1:100]) based on a subset of markers

terms

indices of the coefficients to be tested. The indices corresponding to SNP are used if terms is not specified

rank

decomposition is truncated to the first rank eigen-vectors

map

p x 2 matrix where each entry corresponds to a marker in features. First column is the marker names, second columns is the genetic or physical location

distance

size of the proximal contamination window in units specifed by map.

dcmp_features

the indices in features of the markers used to construct dcmp

W_til

markers used to construct decomp that should now be removed from costruction of decomp. This is the proximal contamination term of Listgarten, et al. (2012)

scale

should W_til be scaled and centered

delta

ratio of variance components governing the fit of the model. This should be estimated from a previous evaluation of `lm' on the same response and eigen-decomposition

reEstimateDelta

should delta be re-estimated for every marker. Note: reEstimateDelta=TRUE is much slower

nthreads

number of to use for parallel execution

verbose

print extra information

progress

show progress bar

cincl

column indeces of features to include for analysis

cexcl

column indeces of features to exclude for analysis

Examples

# Generate data
n = 100
p = 500
X = matrix(sample(0:2, n*p, replace=TRUE), nrow=n)
y = rnorm(n)
sex = as.factor(sample(1:2, n, replace=TRUE))

K = tcrossprod(matrix(rnorm(n*n*3), nrow=n))
decomp = eigen(K, symmetric=TRUE)

# Fit null model
fit = lrgpr( y ~ sex, decomp)

# Fit model for all markers 
pValues = lrgprApply( y ~ sex + sex:SNP, features=X, decomp, terms=c(3,4), delta=fit$delta)
#> 500/500 [=========================] 100.00%    0h 0m 0s