Fit a Low Rank Gaussian Process Regression (LRGPR) / Linear Mixed Model (LMM)

Fit LRGPR/LMM models that account for covariance in response values, but where the scale of the covariance is unknown. Standard linear modeling syntax is used for the model specification in addition to a covariance matrix or its eigen-decomposition.

Usage

lrgpr(
  formula,
  decomp,
  rank = max(ncol(decomp$u), ncol(decomp$vectors)),
  delta = NULL,
  nthreads = 4,
  W_til = NULL,
  scale = TRUE,
  diagnostic = FALSE
)

Arguments

formula

standard linear modeling syntax as used in `lm'

decomp

eigen-decomposition produced from eigen(K), where K is the covariance matrix. Or singular value decomposition svd(X[,1:100]) based on a subset of markers

rank

decomposition is truncated to the first rank eigen-vectors

delta

ratio of variance components governing the fit of the model. This should be estimated from a previous evaluation of `lm' on the same response and eigen-decomposition

nthreads

number of threads to use for parallel execution

W_til

markers used to construct decomp that should now be removed from costruction of decomp. This is the proximal contamination term of Listgarten, et al. (2012)

scale

should W_til be scaled and centered

diagnostic

compute diagnostic statistics to be used with plot()

Value

coefficients

regression coefficients for each covariate

p.values

p-values from Wald test of each coefficient

sd

standard deviation of each coefficient estimate

sigSq_e

variance component $$\sigma^2_e$$ corresponding to the residual error

sigSq_a

variance component $$\sigma^2_a$$ corresponding the scale of the covariance, K

delta

ratio of variance components: $$\sigma^2_e / \sigma^2_a$$

rank

the rank of the random effect

logLik

log-likelihood of the model fit

fitted.values

estimated response values: y_hat

alpha

BLUP of the random effect

Sigma

variance-covariate matrix of estimate of beta

hii

diagonals of the matrix H such that y_hat = Hy

y

responses

x

design matrix

df

effective degrees of freedom: trace(H) based on Hoffman (2013)

residuals

residuals of model fit: y - y_hat

AIC

Akaike information criterion

BIC

Bayesian information criterion

GCV

generalized cross-validation

eigenVectors

eigen-vectors in decomp

eigenValues

eigen-values in decomp

df.residual

n - ncol(X)

rank

rank of decomposition used, where only non-negative eigen/singular values are considered

Details

lrgpr fits the model: $$ y = X \beta + \alpha + \epsilon$$

$$\alpha \sim N(0, K \sigma^2_a)$$

$$\epsilon \sim N(0, \sigma^2_e)$$

where $$\delta = \sigma^2_e / \sigma^2_a$$

In practice the eigen-decomposition of K, and not K itself is required. The rank can be set to use only eigen-vectors 1:rank in the model.

This package allows hypothesis tests of single coefficients using fit$p.values which fits a Wald test. Composite hypothesis tests of multiple coefficients are performed with wald(fit, terms=1:3).

Note that likelihood ratio tests with linear mixed models do not perform well and the resulting p-values often do not follow a uniform distribution under the null (Pinheiro and Bates, 2000). We strongly advise against using it with this model.

lrgpr uses the algorithm of Lippert, et al. (2011).

See Hoffman (2013) for an interpretation of the linear mixed model.

References

Kang, H. M., et al. (2010) Variance component model to account for sample structure in genome-wide association studies. _Nature Genetics_ 42, 348-54

Lippert, C., et al. (2011) FaST linear mixed models for genome-wide association studies. _Nature Methods_ 9, 525-26

Listgarten, J., et al. (2012) Improved linear mixed models for genome-wide association studies. _Nature Methods_ 8, 833-5

Rasmussen, C. E. and Williams, C. K. I. (2006) Gaussian processes for machine learning. MIT Press

Pinheiro, J. C. and Bates, D. M. (2000) Mixed-Effects Models in S and S-PLUS. Springer, New York

Hoffman, G. E. (2013) Correcting for Population Structure and Kinship Using the Linear Mixed Model: Theory and Extensions. _PLoS ONE_ 8(10):e75707

Note that degrees freedom and some diagnostic statistics are not currently calculated when W_til is specified.

See also

Examples

# Generate random data
set.seed(1)
n <- 200
y <- rnorm(n)
K <- crossprod( matrix(rnorm(n*1000), ncol=n) )
age <- rpois(n, 50)
sex <- as.factor(sample(1:2, n, replace=TRUE))
decomp <- eigen(K)

# Fit the model
fit <- lrgpr( y ~ sex + age, decomp, diagnostic=TRUE) 
#> Warning: non-list contrasts argument ignored
#> Error in lrgpr(y ~ sex + age, decomp, diagnostic = TRUE): object 'X' not found

# Print results
fit
#> Error in eval(expr, envir, enclos): object 'fit' not found

# Print more detailed results 
summary(fit)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'summary': object 'fit' not found

# P-values for each covariate
fit$p.values
#> Error in eval(expr, envir, enclos): object 'fit' not found

# Visualize fit of the model like for `lm'
par(mfrow=c(2,2))
plot(fit)
#> Error in eval(expr, envir, enclos): object 'fit' not found

# Composite hypothesis test using Wald's test
# Joint test of coefficients 2:3
wald( fit, terms=2:3)
#> Error in is.null(terms) || is.na(terms): 'length = 2' in coercion to 'logical(1)'